+++ src/components/SideMenu/SideMenuAdmin/index.js
@@ -1,0 +1,3 @@
+import SideMenuAdmin from './SideMenuAdmin';
+
+export default SideMenuAdmin;
