+++ src/components/SideMenu/SideMenuPhotographer/index.js
@@ -1,0 +1,3 @@
+import SideMenuPhotographer from './SideMenuPhotographer';
+
+export default SideMenuPhotographer;
