+++ src/components/SideMenu/SideMenuUser/index.js
@@ -1,0 +1,3 @@
+import SideMenuUser from './SideMenuUser';
+
+export default SideMenuUser;
